Added exist check

+++ project-service/internal/handlers/delete.go
@@ -3,6 +3,7 @@ package handlers
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/lilahamstern/bec-microservices/project-service/internal/database"
+	"github.com/lilahamstern/bec-microservices/project-service/internal/model"
 	"github.com/lilahamstern/bec-microservices/project-service/internal/utils"
 	"net/http"
 )
@@ -12,6 +13,13 @@ func DeleteProject(c *gin.Context) {
 
 	var projectId = c.Param("projectId")
 
+	if exists := db.ProjectExists(model.Project{
+		Id: projectId,
+	}); exists == nil {
+		utils.WriteJsonMessage(c.Writer, http.StatusNotFound, "Project could not be found")
+		return
+	}
+
 	err := db.DeleteProject(projectId)
 
 	if err != nil {
@@ -19,5 +27,5 @@ func DeleteProject(c *gin.Context) {
 		return
 	}
 
-	utils.WriteJsonMessage(c.Writer, http.StatusNotFound, "project delete")
+	utils.WriteJsonMessage(c.Writer, http.StatusAccepted, "project deleted")
 }
